+++ tests/testthat/test-tar_render_manuscript.R
@@ -5,27 +5,34 @@ tar_test("tar_render_manuscript works for qmd input", {
   use_project_directory()
   tar_script({
     qmd_file <- "reports/manuscript.qmd"
-    list(tar_target(table_1, "Table 1"),
-         tar_target(figure_1, "Figure 1"),
-         tar_render_manuscript(manuscript, qmd_file, quiet = FALSE))
+    appendix_qmd_file <- "reports/appendix.qmd"
+    list(tar_render_manuscript(appendix, appendix_qmd_file, appendix = TRUE),
+         tar_render_manuscript(manuscript, qmd_file))
   }, ask = FALSE)
   # manifest
   out <- tar_manifest(callr_function = NULL)
-  expect_equal(nrow(out), 3L)
-  expect_equal(out$name, c("figure_1", "table_1", "manuscript"))
+  expect_equal(nrow(out), 2L)
+  expect_setequal(out$name, c("appendix", "manuscript"))
   # graph
   out <- tar_network(callr_function = NULL, targets_only = TRUE)$edges
-  expect_equal(out, tibble::tibble(from = c("figure_1", "table_1"), to = "manuscript"))
+  expect_equal(out, tibble::tibble(from = c("appendix"), to = "manuscript"))
   # results
   suppressWarnings(tar_make(callr_function = NULL))
+  expect_setequal(fs::path_abs(tar_read(appendix)[[1]]),
+                  fs::path_real(c("reports/appendix.typ")))
+  expect_setequal(
+    fs::path_abs(tar_read(appendix)[2:3]),
+    fs::path_real(c("reports/appendix.qmd", "reports/_setup.qmd"))
+  )
   expect_setequal(fs::path_abs(tar_read(manuscript)[[1]]),
                   fs::path_real(c("output/manuscript.pdf")))
   expect_setequal(
-    fs::path_abs(tar_read(manuscript)[2:3]),
-    fs::path_real(c("reports/manuscript.qmd", "reports/_setup.qmd"))
+    fs::path_abs(tar_read(manuscript)[2:5]),
+    fs::path_real(c("reports/manuscript.qmd", "reports/_setup.qmd",
+                    "reports/_figures/figure-1.qmd", "reports/_tables/table-1.qmd"))
   )
-  expect_true(all(fs::file_exists(tar_read(manuscript)[1:2])))
-  expect_false(any(fs::file_exists(c("reports/manuscript.docx", "reports/manuscript.pdf"))))
+  expect_true(all(fs::file_exists(tar_read(manuscript))))
+  expect_false(any(fs::file_exists(c("reports/manuscript.pdf"))))
   # Everything should be up to date.
   expect_equal(tar_outdated(callr_function = NULL), character(0))
 })